IHostedService - tests
